♻️refactor: UserList, ManagerControl 분리
- page가 너무 커져서 분리
